DRY: extract shared helpers into dedicated lib modules
Extract duplicated functions and constants into shared modules:
- lib/http-utils.ts: encodeBasicAuth, prepareBody, toUserSession
- lib/runtime-utils.ts: client generation helpers, type aliases
- lib/service-config.ts: base URLs, path prefixes, legacy prefixes

Net: -218 lines. Build + 95 tests pass.
