+++ packages/api/src/runtime/create-public-client.ts
@@ -1,22 +1,17 @@
 import type {
-  HttpMethod,
-  OperationDefinition,
-  ServiceOperationCatalog,
-} from "@/generated/public-contracts";
-import type {
   ApiParams,
   OperationRequest,
   PublicClientConfig,
   TokenResponse,
-  UserSession,
 } from "@/types";
 import { publicOperationCatalog } from "@/generated/public-contracts";
+import { toUserSession } from "@/lib/http-utils";
+import { createGeneratedGroups, createRawClient } from "@/lib/runtime-utils";
 import { replacePathParams } from "@/lib/url";
 import { PublicQuranFetcher } from "@/sdk/public-fetcher";
 
 type RawOperation = (request?: OperationRequest) => Promise<unknown>;
 type GeneratedGroup = Record<string, RawOperation>;
-type GeneratedGroups = Record<string, GeneratedGroup>;
 type QuranReflectFacade = {
   comments: GeneratedGroup;
   posts: GeneratedGroup;
@@ -28,135 +23,6 @@ type QuranReflectFacade = {
 
 const SERVER_ONLY_MESSAGE =
   "This API requires @quranjs/api/server or a backend. Content and search are server-side for confidential clients.";
-
-const HTTP_METHOD_TO_MUTATION_NAME: Record<HttpMethod, string> = {
-  delete: "remove",
-  get: "list",
-  patch: "update",
-  post: "create",
-  put: "update",
-};
-
-const toCamelCase = (value: string): string => {
-  const parts = value
-    .replace(/([a-z0-9])([A-Z])/g, "$1 $2")
-    .replace(/[^A-Za-z0-9]+/g, " ")
-    .trim()
-    .split(/\s+/)
-    .filter(Boolean);
-
-  if (parts.length === 0) {
-    return "misc";
-  }
-
-  return parts
-    .map((part, index) => {
-      const lower = part.toLowerCase();
-      if (index === 0) {
-        return lower;
-      }
-
-      return lower.charAt(0).toUpperCase() + lower.slice(1);
-    })
-    .join("");
-};
-
-const toPascalCase = (value: string): string => {
-  const camel = toCamelCase(value);
-  return camel.charAt(0).toUpperCase() + camel.slice(1);
-};
-
-const isPathParam = (segment: string): boolean =>
-  segment.startsWith("{") && segment.endsWith("}");
-
-const getResourceSegments = (path: string): string[] => {
-  const segments = path.split("/").filter(Boolean);
-  if (segments[0]?.match(/^v\d+$/)) {
-    return segments.slice(1);
-  }
-
-  return segments;
-};
-
-const pickGeneratedMethodName = (operation: OperationDefinition): string => {
-  const resourceSegments = getResourceSegments(operation.path);
-  const lastSegment = resourceSegments.at(-1);
-  const literalSegments = resourceSegments.filter(
-    (segment) => !isPathParam(segment),
-  );
-  const lastLiteral = literalSegments.at(-1);
-  const groupLiteral = literalSegments[0];
-
-  if (resourceSegments.length === 1) {
-    return HTTP_METHOD_TO_MUTATION_NAME[operation.method];
-  }
-
-  if (
-    resourceSegments.length === 2 &&
-    lastSegment &&
-    isPathParam(lastSegment)
-  ) {
-    if (operation.method === "get") {
-      return "get";
-    }
-
-    if (operation.method === "delete") {
-      return "remove";
-    }
-
-    return "update";
-  }
-
-  if (lastSegment && !isPathParam(lastSegment) && lastLiteral) {
-    if (lastLiteral === groupLiteral) {
-      return `${HTTP_METHOD_TO_MUTATION_NAME[operation.method]}ById`;
-    }
-
-    return toCamelCase(lastLiteral);
-  }
-
-  return toCamelCase(operation.operationName);
-};
-
-const createRawClient = (
-  fetcher: PublicQuranFetcher,
-  section: ServiceOperationCatalog,
-): Record<string, RawOperation> => {
-  const rawOperations: Record<string, RawOperation> = {};
-
-  for (const [operationName, operation] of Object.entries(section.operations)) {
-    rawOperations[operationName] = (request?: OperationRequest) =>
-      fetcher.requestOperation(operation, request);
-  }
-
-  return rawOperations;
-};
-
-const createGeneratedGroups = (
-  section: ServiceOperationCatalog,
-  rawOperations: Record<string, RawOperation>,
-): GeneratedGroups => {
-  const groups: GeneratedGroups = {};
-
-  for (const [operationName, operation] of Object.entries(section.operations)) {
-    const literals = getResourceSegments(operation.path).filter(
-      (segment) => !isPathParam(segment),
-    );
-    const groupName = toCamelCase(literals[0] ?? "misc");
-    const suggestedName = pickGeneratedMethodName(operation);
-    const currentGroup = groups[groupName] ?? {};
-    let methodName = suggestedName;
-
-    if (currentGroup[methodName]) {
-      methodName = `${suggestedName}${toPascalCase(operationName)}`;
-    }
-
-    currentGroup[methodName] = rawOperations[operationName]!;
-    groups[groupName] = currentGroup;
-  }
-
-  return groups;
-};
 
 const createUserServiceRequest =
   (fetcher: PublicQuranFetcher, service: "auth" | "quranReflect") =>
@@ -322,37 +188,6 @@ const createQuranReflectFacade = (
     rooms: generatedGroups.rooms ?? {},
     tags: generatedGroups.tags ?? {},
     users: generatedGroups.users ?? {},
-  };
-};
-
-const toUserSession = (
-  token: TokenResponse & {
-    accessToken?: string;
-    expiresIn?: number;
-    idToken?: string;
-    refreshToken?: string;
-    tokenType?: string;
-  },
-  currentSession?: Partial<UserSession> | null,
-): UserSession => {
-  const expiresIn = token.expiresIn ?? token.expires_in;
-  const accessToken = token.accessToken ?? token.access_token;
-  const idToken = token.idToken ?? token.id_token ?? currentSession?.idToken;
-  const refreshToken =
-    token.refreshToken ?? token.refresh_token ?? currentSession?.refreshToken;
-  const tokenType =
-    token.tokenType ?? token.token_type ?? currentSession?.tokenType;
-  const expiresAt = expiresIn
-    ? Date.now() + expiresIn * 1000
-    : currentSession?.expiresAt;
-
-  return {
-    accessToken,
-    expiresAt,
-    idToken,
-    refreshToken,
-    scope: token.scope ?? currentSession?.scope,
-    tokenType,
   };
 };
